requirements/base.txt
celery~=5.2
Django~=4.1
django-cors-headers~=3.1
django-extensions~=3.2.0
django-filter~=22.1
django-model-utils~=4.2
djangorestframework~=3.13
djangorestframework-simplejwt~=5.2
drf-spectacular~=0.23
drf-spectacular-sidecar~=2022.8
mixer~=7.2
Pillow~=9.2
selenium~=4.5
user_agent~=0.1.10
